requirements.txt
fastapi
uvicorn
pdfplumber
python-dotenv
anthropic
httpx
python-multipart
